dto修改

+++ src/main/java/org/csu/dto/LocationCreateDto.java
@@ -3,6 +3,8 @@ package org.csu.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class LocationCreateDto {
 
@@ -10,10 +12,10 @@ public class LocationCreateDto {
     private Long herbId;
 
     @NotNull(message = "经度不能为空")
-    private Double longitude;
+    private BigDecimal longitude;
 
     @NotNull(message = "纬度不能为空")
-    private Double latitude;
+    private BigDecimal latitude;
 
     @NotNull(message = "省份不能为空")
     private String province;
